pyproject.toml
[tool.poetry]
name = "drf-lms"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
python-dotenv = "^1.0.1"
django = "4.x"
djangorestframework = "^3.15.2"
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
